Remove Haskell display parts

+++ pallas-network/src/miniprotocols/localstate/queries_v16/mod.rs
@@ -265,7 +265,7 @@ pub struct ExUnitPrices {
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct ExUnits {
     #[n(0)]
-    pub mem: u32,
+    pub mem: u64,
     #[n(1)]
     pub steps: u64,
 }
